Move all sidebar logic to collection's route logic
